requirements.txt
absl-py==0.8.1
asn1crypto==1.2.0
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
dill==0.3.1.1
future==0.18.2
gast==0.2.2
google-pasta==0.1.8
googleapis-common-protos==1.6.0
grpcio==1.16.1
h5py==2.9.0
idna==2.8
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.4
olefile==0.46
opt-einsum==3.1.0
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
promise==2.2.1
prompt-toolkit==3.0.2
protobuf==3.11.2
psutil==5.6.7
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.5.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
PySocks==1.7.1
pyspark==2.4.3
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.0
requests==2.22.0
scipy==1.3.2
sip==4.19.13
six==1.13.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-datasets==1.2.0
tensorflow-estimator==2.0.0
tensorflow-hub==0.7.0
tensorflow-metadata==0.14.0
termcolor==1.1.0
tornado==6.0.3
tqdm==4.40.2
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
